requirements.txt
requests==2.31.0
pandas==2.1.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.1
pytest==7.4.3
streamlit==1.28.2
plotly==5.18.0
